refactor :: folder structure
