+++ src/screens/main/search/components/searchBar/SearchBar.style.ts
